requirements.txt
requests
bs4
flask
flask-RESTful
db-sqlite3
toml
pyyaml
bcrypt